+++ cassandra/io/libevreactor.py
@@ -14,7 +14,16 @@ from cassandra.connection import (Connection, ResponseWaiter, ConnectionShutdown
                                   MAX_STREAM_PER_CONNECTION)
 from cassandra.decoder import RegisterMessage
 from cassandra.marshal import int32_unpack
-import cassandra.io.libevwrapper as libev
+try:
+    import cassandra.io.libevwrapper as libev
+except ImportError:
+    raise ImportError(
+        "The C extension needed to use libev was not found.  This "
+        "probably means that you didn't have the required build dependencies "
+        "when installing the driver.  See "
+        "http://datastax.github.io/python-driver/installation.html#c-extensions "
+        "for instructions on installing build dependencies and building "
+        "the C extension.")
 
 try:
     from cStringIO import StringIO