+++ CHANGELOG.rst
@@ -20,6 +20,7 @@ Other
 -----
 * Don't ignore column names when parsing typestrings.  This is needed for
   user-defined type support.  (github issue #90)
+* Better error message when libevwrapper is not found
 
 1.0.2
 =====